@smacker 2019-03-19
thanks! I'm merging it for now and I'll add 2 more columns for annotated & semantic modes later.
(imo both cases, "fair" comparison & overhead to translating it to annotated/semantic, are interesting)